Fixes Style CI with rule to skip specifically named tests.
